local/clients/cli/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ucp-chat"
version = "0.1.0"
description = "Universal Context Protocol CLI - Multi-provider LLM chat with intelligent tool injection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "UCP Team"}
]
keywords = ["llm", "mcp", "ai", "tools", "context"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiofiles>=24.0",
    "prompt-toolkit>=3.0",
    "keyring>=25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "black>=24.0",
    "ruff>=0.5",
]

[project.scripts]
ucp-chat = "ucp_chat.cli:app"

[project.urls]
Homepage = "https://github.com/ucp/ucp-chat"
Documentation = "https://github.com/ucp/ucp-chat#readme"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
